+++ metrics-util/src/registry/mod.rs
@@ -1,0 +1,13 @@
+//! High-performance metrics storage.
+
+mod storage;
+pub use storage::{Storage, AtomicStorage};
+
+#[cfg(feature = "recency")]
+mod recency;
+
+#[cfg(feature = "recency")]
+pub use recency::{Generation, GenerationalAtomicStorage, Recency};
+
+mod registry;
+pub use registry::Registry;
